+++ ocrd_kraken/segment.py
@@ -17,6 +17,7 @@ from ocrd_modelfactory import page_from_file
 
 import shapely.geometry as geom
 from shapely.prepared import prep as geom_prep
+import torch
 
 from .config import OCRD_TOOL
 
@@ -50,7 +51,12 @@ class KrakenSegment(Processor):
             log.info("Using blla segmenter")
             blla_model_fname = self.resolve_resource(self.parameter['blla_model'])
             kwargs['model'] = TorchVGSLModel.load_model(blla_model_fname)
-            kwargs['device'] = self.parameter['device']
+            device = self.parameter['device']
+            if device != 'cpu' and not torch.cuda.is_available():
+                device = 'cpu'
+            if device == 'cpu':
+                log.warning("no CUDA device available. Running without GPU will be slow")
+            kwargs['device'] = device
         def segmenter(img):
             return segment(img, **kwargs)
         self.segmenter = segmenter
@@ -118,13 +124,16 @@ class KrakenSegment(Processor):
                     for idx_line, line_dict in enumerate(res['lines']):
                         line_poly = coordinates_for_segment(line_dict['boundary'], None, page_coords)
                         line_baseline = coordinates_for_segment(line_dict['baseline'], None, page_coords)
+                        line_id = 'region_%s_line_%s' % (idx_region + 1, idx_line + 1)
+                        line_type = line_dict.get('tags', {}).get('type', '')
+                        log.info("Line %s is of type %s", line_id, line_type)
                         line_polygon = make_valid(geom.Polygon(line_poly))
                         if region_polygon.contains(line_polygon):
                             if idx_line in handled_lines:
                                 log.error("Line %s was already added to region %s" % (idx_line, handled_lines[idx_line]))
                                 continue
                             region_elem.add_TextLine(TextLineType(
-                                id='region_%s_line_%s' % (idx_region + 1, idx_line + 1),
+                                id=line_id,
                                 Baseline=BaselineType(points=points_from_polygon(line_baseline)),
                                 Coords=CoordsType(points=points_from_polygon(line_poly))))
                             handled_lines[idx_line] = idx_region